memory load/store logic moved to executor
