+++ executor/wrapper/wrapperInstance.go
@@ -91,6 +91,21 @@ func (inst *WrapperInstance) GetMemory() executor.Memory {
 	return inst.wrappedInstance.GetMemory()
 }
 
+// MemLoad returns the contents from the given offset of the WASM memory.
+func (inst *WrapperInstance) MemLoad(offset int32, length int32) ([]byte, error) {
+	return inst.wrappedInstance.MemLoad(offset, length)
+}
+
+// MemLoadMultiple returns multiple byte slices loaded from the WASM memory, starting at the given offset and having the provided lengths.
+func (inst *WrapperInstance) MemLoadMultiple(offset int32, lengths []int32) ([][]byte, error) {
+	return inst.wrappedInstance.MemLoadMultiple(offset, lengths)
+}
+
+// MemStore stores the given data in the WASM memory at the given offset.
+func (inst *WrapperInstance) MemStore(offset int32, data []byte) error {
+	return inst.wrappedInstance.MemStore(offset, data)
+}
+
 // IsFunctionImported wraps the call to the underlying instance.
 func (inst *WrapperInstance) IsFunctionImported(name string) bool {
 	result := inst.wrappedInstance.IsFunctionImported(name)
